+++ src/task/task.ts
@@ -7,7 +7,7 @@ import type {Filesystem} from 'src/fs/filesystem.js';
 export interface Task<TArgs = Args, TEvents = {}> {
 	run: (ctx: TaskContext<TArgs, TEvents>) => Promise<unknown>; // TODO return value (make generic, forward it..how?)
 	summary?: string;
-	dev?: boolean;
+	production?: boolean;
 }
 
 export interface TaskContext<TArgs = {}, TEvents = {}> {
@@ -21,7 +21,6 @@ export interface TaskContext<TArgs = {}, TEvents = {}> {
 		taskName: string,
 		args?: Args,
 		events?: StrictEventEmitter<EventEmitter, TEvents>,
-		dev?: boolean,
 		fs?: Filesystem,
 	) => Promise<void>;
 }
@@ -51,3 +50,17 @@ export interface Args {
 	_: string[];
 	[key: string]: unknown; // can assign anything to `args` in tasks
 }
+
+export const serializeArgs = (args: Args): string[] => {
+	const result: string[] = [];
+	let _: string[] | null = null;
+	for (const [key, value] of Object.entries(args)) {
+		if (key === '_') {
+			_ = (value as any[]).map((v) => v.toString());
+		} else {
+			result.push(`--${key}`);
+			result.push((value as any).toString());
+		}
+	}
+	return _ ? [...result, ..._] : result;
+};